Y3-Sem2/AIシステム開発演習I/第14週_ディープラーニングのWebAPI/server14-2_SK3A04文家俊/requirements.txt
hug
pillow
joblib==1.4.2